requirements.txt
setuptools~=52.0.0
pandas~=1.2.4
yaml~=0.2.5
pyyaml~=5.4.1
sqlalchemy~=1.4.15
xarray~=0.17.0
gym~=0.18.0
numpy~=1.20.2
scipy~=1.6.2
python-binance
